examples/todo_hive/worker_a_src/requirements.txt
flask==3.0.0
bcrypt==4.0.1
PyJWT==2.8.0